+++ packages/appkit/src/client.ts
@@ -215,7 +215,6 @@ export class AppKit {
   }
 
   private async initialize(options: AppKitOptionsWithSdk) {
-    console.log('###?? initialize : start ', new Date().toLocaleTimeString())
     this.initControllers(options)
     await this.initChainAdapters()
     await this.injectModalUi()
@@ -242,7 +241,6 @@ export class AppKit {
 
   // -- Public -------------------------------------------------------------------
   public async open(options?: OpenOptions) {
-    console.log('###?? open : start ', new Date().toLocaleTimeString())
     await this.injectModalUi()
     if (options?.uri && this.universalProvider) {
       ConnectionController.setUri(options.uri)
@@ -256,7 +254,6 @@ export class AppKit {
   }
 
   public async close() {
-    console.log('###?? close : start ', new Date().toLocaleTimeString())
     await this.injectModalUi()
     ModalController.close()
   }
@@ -275,7 +272,6 @@ export class AppKit {
   }
 
   public async switchNetwork(appKitNetwork: AppKitNetwork) {
-    console.log('###?? switchNetwork : start ', new Date().toLocaleTimeString())
     const network = this.caipNetworks?.find(n => n.id === appKitNetwork.id)
 
     if (!network) {
@@ -685,7 +681,6 @@ export class AppKit {
   }
 
   public async disconnect() {
-    console.log('###?? disconnect : start ', new Date().toLocaleTimeString())
     await ConnectionController.disconnect()
   }
 
@@ -888,7 +883,6 @@ export class AppKit {
   }
 
   private async initializeBlockchainApiController() {
-    console.log('###?? initializeBlockchainApiController : start ', new Date().toLocaleTimeString())
     /*
      * TODO: implement api
      * await BlockchainApiController.getSupportedNetworks()
@@ -962,8 +956,6 @@ export class AppKit {
   private createClients() {
     this.connectionControllerClient = {
       connectWalletConnect: async () => {
-        console.log('###?? connectWalletConnect : start ', new Date().toLocaleTimeString())
-
         const adapter = this.getAdapter(ChainController.state.activeChain)
 
         if (!adapter) {
@@ -978,7 +970,6 @@ export class AppKit {
         await this.syncWalletConnectAccount()
       },
       connectExternal: async ({ id, info, type, provider, chain, caipNetwork }) => {
-        console.log('###?? connectExternal : start ', new Date().toLocaleTimeString())
         console.log(`connectExternal`)
 
         const activeChain = ChainController.state.activeChain as ChainNamespace
@@ -1020,7 +1011,6 @@ export class AppKit {
         this.setAllAccounts(accounts, chainToUse)
       },
       reconnectExternal: async ({ id, info, type, provider }) => {
-        console.log('###?? reconnectExternal : start ', new Date().toLocaleTimeString())
         const namespace = ChainController.state.activeChain as ChainNamespace
         const adapter = this.getAdapter(namespace)
         if (adapter?.reconnect) {
@@ -1029,7 +1019,6 @@ export class AppKit {
         }
       },
       disconnect: async () => {
-        console.log('###?? disconnect : start ', new Date().toLocaleTimeString())
         const namespace = ChainController.state.activeChain as ChainNamespace
         const adapter = this.getAdapter(namespace)
         const provider = ProviderUtil.getProvider(namespace)
@@ -1054,7 +1043,6 @@ export class AppKit {
         return ids.some(id => Boolean(window.ethereum?.[String(id)]))
       },
       signMessage: async (params: { message: string; customData?: CustomData }) => {
-        console.log('###?? signMessage : start appkit : ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
         const result = await adapter?.signMessage({
           message: params.message,
@@ -1065,8 +1053,17 @@ export class AppKit {
 
         return result?.signature || ''
       },
+      etherSignMessage: async (params: { message: string; address: `0x${string}` }) => {
+        const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
+        const result = await adapter?.etherSignMessage({
+          message: params.message,
+          address: AccountController.state.address as string,
+          provider: ProviderUtil.getProvider(ChainController.state.activeChain as ChainNamespace)
+        })
+
+        return result?.signature || ''
+      },
       signEIP712: async (args: SignEIP712Args) => {
-        console.log('###?? signEIP712 : start ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
 
         const result = await adapter?.signEIP712({
@@ -1088,7 +1085,6 @@ export class AppKit {
         return result?.signature || ''
       },
       sendTransaction: async (args: SendTransactionArgs) => {
-        console.log('###?? sendTransaction : start ', new Date().toLocaleTimeString())
         if (args.chainNamespace === ConstantsUtil.CHAIN.EVM) {
           const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
 
@@ -1106,7 +1102,6 @@ export class AppKit {
         return null
       },
       estimateGas: async (args: EstimateGasTransactionArgs) => {
-        console.log('###?? estimateGas : start ', new Date().toLocaleTimeString())
         if (args.chainNamespace === ConstantsUtil.CHAIN.EVM) {
           const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
           const provider = ProviderUtil.getProvider(
@@ -1129,7 +1124,6 @@ export class AppKit {
         return 0n
       },
       getEnsAvatar: async () => {
-        console.log('###?? getEnsAvatar : start ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
         const result = await adapter?.getProfile({
           address: AccountController.state.address as string,
@@ -1139,7 +1133,6 @@ export class AppKit {
         return result?.profileImage || false
       },
       getEnsAddress: async (name: string) => {
-        console.log('###?? getEnsAddress : start ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
         const caipNetwork = this.getCaipNetwork()
         if (!caipNetwork) {
@@ -1153,7 +1146,6 @@ export class AppKit {
         return result?.address || false
       },
       writeContract: async (args: WriteContractArgs) => {
-        console.log('###?? writeContract : start ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
         const caipNetwork = this.getCaipNetwork()
         const caipAddress = this.getCaipAddress()
@@ -1169,7 +1161,6 @@ export class AppKit {
         return result ? { hash: result?.hash as `0x${string}` } : null
       },
       readContract: async (args: ReadContractArgs) => {
-        console.log('###?? readContract : start ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
         const caipNetwork = this.getCaipNetwork()
         const provider = ProviderUtil.getProvider(
@@ -1198,19 +1189,16 @@ export class AppKit {
         return adapter?.formatUnits({ value, decimals }) ?? '0'
       },
       getCapabilities: async (params: AdapterBlueprint.GetCapabilitiesParams) => {
-        console.log('###?? getCapabilities : start ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
 
         return await adapter?.getCapabilities(params)
       },
       grantPermissions: async (params: AdapterBlueprint.GrantPermissionsParams) => {
-        console.log('###?? grantPermissions : start ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
 
         return await adapter?.grantPermissions(params)
       },
       revokePermissions: async (params: AdapterBlueprint.RevokePermissionsParams) => {
-        console.log('###?? revokePermissions : start ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
 
         if (adapter?.revokePermissions) {
@@ -1220,7 +1208,6 @@ export class AppKit {
         return '0x'
       },
       walletGetAssets: async (params: AdapterBlueprint.WalletGetAssetsParams) => {
-        console.log('###?? walletGetAssets : start ', new Date().toLocaleTimeString())
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
 
         return (await adapter?.walletGetAssets(params)) ?? {}
@@ -1229,7 +1216,6 @@ export class AppKit {
 
     this.networkControllerClient = {
       switchCaipNetwork: async caipNetwork => {
-        console.log('###?? switchCaipNetwork : start ', new Date().toLocaleTimeString())
         if (!caipNetwork) {
           return
         }
@@ -1437,7 +1423,6 @@ export class AppKit {
   }
 
   private async syncAuthConnector(provider: W3mFrameProvider) {
-    console.log('###?? syncAuthConnector : start ', new Date().toLocaleTimeString())
     this.setLoading(true)
     const isLoginEmailUsed = provider.getLoginEmailUsed()
     this.setLoading(isLoginEmailUsed)
@@ -1527,10 +1512,6 @@ export class AppKit {
       })
 
       this.universalProvider.on('session_event', (callbackData: unknown) => {
-        console.log(
-          'this.universalProvider ###?? session_event : start ',
-          new Date().toLocaleTimeString()
-        )
         if (WcHelpersUtil.isSessionEventData(callbackData)) {
           const { name, data } = callbackData.params.event
 
@@ -1624,7 +1605,6 @@ export class AppKit {
   }
 
   private async updateNativeBalance(ignoreCache = false) {
-    console.log('###?? updateNativeBalance : start ', new Date().toLocaleTimeString())
     const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
     if (adapter && ChainController.state.activeChain && AccountController.state.address) {
       const balance = await adapter.getBalance({
@@ -1655,7 +1635,6 @@ export class AppKit {
   }
 
   private async syncWalletConnectAccount() {
-    console.log('###?? syncWalletConnectAccount : start ', new Date().toLocaleTimeString())
     const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
 
     this.chainNamespaces.forEach(async chainNamespace => {
@@ -1757,7 +1736,6 @@ export class AppKit {
       chainNamespace: ChainNamespace
     }
   ) {
-    console.log('###?? syncAccount : start ', new Date().toLocaleTimeString())
     const { address, chainId, chainNamespace } = params
 
     const { chainId: activeChainId } = StorageUtil.getActiveNetworkProps()
@@ -1834,7 +1812,6 @@ export class AppKit {
     chainId: string | number | undefined
     chainNamespace: ChainNamespace
   }) {
-    console.log('###?? syncBalance : start ', new Date().toLocaleTimeString())
     const caipNetwork = NetworkUtil.getNetworksByNamespace(
       this.caipNetworks,
       params.chainNamespace
@@ -1908,7 +1885,6 @@ export class AppKit {
   }: Pick<AdapterBlueprint.ConnectResult, 'address' | 'chainId'> & {
     chainNamespace: ChainNamespace
   }) {
-    console.log('###?? syncIdentity : start ', new Date().toLocaleTimeString())
     const caipNetworkId: CaipNetworkId = `${chainNamespace}:${chainId}`
     const activeCaipNetwork = this.caipNetworks?.find(n => n.caipNetworkId === caipNetworkId)
 
@@ -1954,7 +1930,6 @@ export class AppKit {
   }
 
   private async syncReownName(address: string, chainNamespace: ChainNamespace) {
-    console.log('###?? syncReownName : start ', new Date().toLocaleTimeString())
     try {
       const registeredWcNames = await this.getReownName(address)
       if (registeredWcNames[0]) {
@@ -1969,7 +1944,6 @@ export class AppKit {
   }
 
   private async syncAdapterConnection(namespace: ChainNamespace) {
-    console.log('###?? syncAdapterConnection : start ', new Date().toLocaleTimeString())
     const adapter = this.getAdapter(namespace)
     const connectorId = StorageUtil.getConnectedConnectorId(namespace)
     const caipNetwork = this.getCaipNetwork()
@@ -2014,7 +1988,6 @@ export class AppKit {
   }
 
   private async syncNamespaceConnection(namespace: ChainNamespace) {
-    console.log('###?? syncNamespaceConnection : start ', new Date().toLocaleTimeString())
     try {
       const connectorId = StorageUtil.getConnectedConnectorId(namespace)
       const isEmailUsed = this.authProvider?.getLoginEmailUsed()
@@ -2042,7 +2015,6 @@ export class AppKit {
   }
 
   private async syncExistingConnection() {
-    console.log('###?? syncExistingConnection : start ', new Date().toLocaleTimeString())
     await Promise.allSettled(
       this.chainNamespaces.map(namespace => this.syncNamespaceConnection(namespace))
     )
@@ -2089,7 +2061,6 @@ export class AppKit {
   }
 
   private async initializeUniversalAdapter() {
-    console.log('###?? initializeUniversalAdapter : start ', new Date().toLocaleTimeString())
     const logger = LoggerUtil.createLogger((error, ...args) => {
       if (error) {
         this.handleAlertError(error)
@@ -2130,7 +2101,6 @@ export class AppKit {
   }
 
   public async getUniversalProvider() {
-    console.log('###?? getUniversalProvider : start ', new Date().toLocaleTimeString())
     if (!this.universalProvider) {
       try {
         await this.createUniversalProvider()
@@ -2174,7 +2144,6 @@ export class AppKit {
   }
 
   private async createUniversalProviderForAdapter(chainNamespace: ChainNamespace) {
-    console.log('###?? createUniversalProviderForAdapter : start ', new Date().toLocaleTimeString())
     await this.getUniversalProvider()
 
     if (this.universalProvider) {
@@ -2252,7 +2221,6 @@ export class AppKit {
   }
 
   private async initChainAdapter(namespace: ChainNamespace) {
-    console.log('###?? initChainAdapter : start ', new Date().toLocaleTimeString())
     this.onConnectors(namespace)
     this.listenAdapter(namespace)
     this.chainAdapters?.[namespace].syncConnectors(this.options, this)
@@ -2261,7 +2229,6 @@ export class AppKit {
   }
 
   private async initChainAdapters() {
-    console.log('###?? initChainAdapters : start ', new Date().toLocaleTimeString())
     await Promise.all(
       this.chainNamespaces.map(async namespace => {
         await this.initChainAdapter(namespace)
@@ -2314,7 +2281,6 @@ export class AppKit {
   }
 
   private async injectModalUi() {
-    console.log('###?? injectModalUi : start ', new Date().toLocaleTimeString())
     if (!this.initPromise && !isInitialized && CoreHelperUtil.isClient()) {
       isInitialized = true
       this.initPromise = new Promise<void>(async resolve => {
@@ -2334,7 +2300,6 @@ export class AppKit {
   }
 
   private async checkExistingSocialConnection() {
-    console.log('###?? checkExistingSocialConnection : start ', new Date().toLocaleTimeString())
     try {
       if (!CoreHelperUtil.isTelegram()) {
         return