+++ packages/appkit/src/networks/bsc/bscMainnet.ts
@@ -2,11 +2,11 @@ import { defineChain } from '../utils.js'
 
 export const bscMainnet = defineChain({
   id: 56,
-  name: 'BNB Smart Chain',
+  name: 'BSC Mainnet',
   nativeCurrency: {
     decimals: 18,
     name: 'BNB',
-    symbol: 'BNB',
+    symbol: 'BNB'
   },
   network: 'bnb',
   rpcUrls: {
@@ -18,14 +18,14 @@ export const bscMainnet = defineChain({
     default: {
       name: 'BscScan',
       url: 'https://bscscan.com',
-      apiUrl: 'https://api.bscscan.com/api',
-    },
+      apiUrl: 'https://api.bscscan.com/api'
+    }
   },
   contracts: {
     multicall3: {
       address: '0xca11bde05977b3631167028862be2a173976ca11',
-      blockCreated: 15921452,
-    },
+      blockCreated: 15921452
+    }
   },
   testnet: false,
   chainNamespace: 'eip155',