+++ packages/appkit/src/networks/bsc/bscTestnet.ts
@@ -2,11 +2,11 @@ import { defineChain } from '../utils.js'
 
 export const bscTestnet = defineChain({
   id: 97,
-  name: 'Binance Smart Chain Testnet',
+  name: 'BSC Testnet',
   nativeCurrency: {
     decimals: 18,
     name: 'BNB',
-    symbol: 'tBNB',
+    symbol: 'tBNB'
   },
   network: 'bsc-testnet',
   rpcUrls: {
@@ -18,14 +18,14 @@ export const bscTestnet = defineChain({
     default: {
       name: 'BscScan',
       url: 'https://testnet.bscscan.com',
-      apiUrl: 'https://api-testnet.bscscan.com/api',
-    },
+      apiUrl: 'https://api-testnet.bscscan.com/api'
+    }
   },
   contracts: {
     multicall3: {
       address: '0xca11bde05977b3631167028862be2a173976ca11',
-      blockCreated: 17422483,
-    },
+      blockCreated: 17422483
+    }
   },
   testnet: true,
   chainNamespace: 'eip155',