+++ examples/sdk-vanilla/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @to-nexus/sdk-example-html
 
+## 1.12.1
+
+### Patch Changes
+
+- @to-nexus/sdk@1.13.0
+
 ## 1.12.0
 
 ### Minor Changes