+++ packages/adapters/bitcoin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @reown/appkit-adapter-bitcoin
 
+## 1.13.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @to-nexus/appkit@1.13.0
+  - @to-nexus/appkit-core@1.13.0
+  - @to-nexus/appkit-utils@1.13.0
+
 ## 1.12.0
 
 ### Minor Changes