+++ packages/adapters/solana/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @reown/appkit-adapter-solana
 
+## 1.13.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @to-nexus/appkit@1.13.0
+  - @to-nexus/appkit-wallet@1.13.0
+  - @to-nexus/appkit-core@1.13.0
+  - @to-nexus/appkit-utils@1.13.0
+  - @to-nexus/appkit-scaffold-ui@1.13.0
+  - @to-nexus/appkit-ui@1.11.0
+
 ## 1.12.0
 
 ### Minor Changes