+++ packages/appkit/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @to-nexus/appkit
 
+## 1.13.0
+
+### Minor Changes
+
+- add eth sign method
+
+### Patch Changes
+
+- Updated dependencies
+  - @to-nexus/appkit-wallet@1.13.0
+  - @to-nexus/appkit-core@1.13.0
+  - @to-nexus/appkit-utils@1.13.0
+  - @to-nexus/appkit-scaffold-ui@1.13.0
+  - @to-nexus/appkit-ui@1.11.0
+  - @to-nexus/appkit-wallet-button@1.13.0
+
 ## 1.12.0
 
 ### Minor Changes