+++ packages/cdn/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @reown/appkit-cdn
 
+## 1.13.0
+
+### Patch Changes
+
+- @to-nexus/sdk@1.13.0
+
 ## 1.12.0
 
 ### Minor Changes