+++ packages/experimental/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @reown/appkit-experimental
 
+## 1.13.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @to-nexus/appkit@1.13.0
+  - @to-nexus/appkit-core@1.13.0
+  - @to-nexus/appkit-ui@1.11.0
+
 ## 1.12.0
 
 ### Patch Changes