+++ packages/siwx/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @reown/appkit-siwx
 
+## 1.13.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @to-nexus/appkit-core@1.13.0
+  - @to-nexus/appkit-utils@1.13.0
+
 ## 1.12.0
 
 ### Patch Changes